painting but not perfect, still no gaming

+++ so_long.h
@@ -18,8 +18,8 @@
 # include <stdio.h>
 # include <fcntl.h>
 # include "Libft/libft.h"
-# include <mlx.h>
-//# include "minilibx_linux/mlx.h"
+//# include <mlx.h>
+# include "minilibx_linux/mlx.h"
 
 
 typedef struct s_map
@@ -47,7 +47,7 @@ typedef struct s_game
 t_map *read_map(int fd, t_map *map);
 t_map *save_map(int fd, t_map *map);
 void get_images(t_game *game);
-void put_image(char c, t_game *game, int i, int j);
+void put_image(char c, t_game *game, size_t i, size_t j);
 void draw_map(t_map *map, t_game *game);
 void initialize_game(t_map *map);
 #endif
